refactor sprite assembler structure
